# Add detail summary to perf report