--- a/src/program.cpp
+++ b/src/program.cpp
@@ -808,12 +808,27 @@ double common_average(const std::vector<double>& v)
     return total / std::distance(v.begin() + n, v.end() - n);
 }
 
-std::string perf_group(const operation& op)
+std::string perf_group(instruction_ref ins, bool detailed)
 {
-    auto attr = op.attributes();
+    std::string result;
+    auto attr = ins->get_operator().attributes();
     if(attr.contains("group"))
-        return attr.at("group").to<std::string>();
-    return op.name();
+        result = attr.at("group").to<std::string>();
+    else
+        result = ins->get_operator().name();
+    if(detailed)
+    {
+        result += "<" + ins->get_shape().type_string();
+        std::vector<std::string> sizes;
+        std::transform(ins->inputs().begin(), ins->inputs().end(), std::back_inserter(sizes), [&](instruction_ref input) {
+            std::string r = to_string_range(input->get_shape().lens(), "x");
+            if(not input->get_shape().standard())
+                r += ":" + to_string_range(input->get_shape().strides(), "x");
+            return r;
+        });
+        result += "(" + join_strings(sizes, ", ") + ")>";
+    }
+    return result;
 }
 
 void program::mark(const parameter_map& params, marker&& m)
@@ -837,7 +852,7 @@ void program::mark(const parameter_map& params, marker&& m)
 void program::perf_report(std::ostream& os,
                           std::size_t n,
                           parameter_map params,
-                          std::size_t batch) const
+                          std::size_t batch, bool detailed) const
 {
     auto& ctx = this->impl->contexts;
     // Run once by itself
@@ -893,9 +908,9 @@ void program::perf_report(std::ostream& os,
     for(auto&& p : ins_vec)
     {
         double avg = common_average(p.second);
-        op_times[perf_group(p.first->get_operator())] += avg;
+        op_times[perf_group(p.first, detailed)] += avg;
         total_instruction_time += avg;
-        op_n[perf_group(p.first->get_operator())]++;
+        op_n[perf_group(p.first, detailed)]++;
     }
     double calculate_overhead_time    = total_time - total_instruction_time;
     double calculate_overhead_percent = calculate_overhead_time * 100.0 / total_time;