With the `--detailed` flag it will add the sizes to the group name. Here is an example output from resnet50:

```
Summary:
gpu::convolution<float_type(1x256x14x14, 256x256x3x3, 2760704, 1x256x14x14)>: 0.332146ms / 5 = 0.0664292ms, 11%
gpu::code_object::mlir_convolution_add_relu<float_type(1x256x14x14:0x1x0x0, 1x1024x14x14, 256x1024x1x1, 1x256x14x14)>: 0.182762ms / 5 = 0.0365524ms, 7%
gpu::convolution<float_type(1x512x7x7, 512x512x3x3, 903168, 1x512x7x7)>: 0.177803ms / 2 = 0.0889014ms, 6%
gpu::convolution<float_type(1x128x28x28, 128x128x3x3, 0, 1x128x28x28)>: 0.176263ms / 3 = 0.0587542ms, 6%
gpu::convolution<float_type(1x64x56x56, 64x64x3x3, 0, 1x64x56x56)>: 0.152718ms / 3 = 0.050906ms, 6%
gpu::code_object::add_relu_kernel<float_type(1x256x14x14, 1x256x14x14:0x1x0x0, 1x256x14x14)>: 0.145661ms / 6 = 0.0242769ms, 5%
gpu::code_object::mlir_convolution_add_add_relu<float_type(1x1024x14x14:0x1x0x0, 1x1024x14x14, 1x256x14x14, 1024x256x1x1, 1x1024x14x14)>: 0.141009ms / 5 = 0.0282018ms, 5%
gpu::code_object::mlir_convolution_add_relu<float_type(1x512x7x7:0x1x0x0, 1x2048x7x7, 512x2048x1x1, 1x512x7x7)>: 0.121047ms / 2 = 0.0605234ms, 5%
gpu::convolution<float_type(1x512x14x14, 512x512x3x3, 903168, 1x512x7x7)>: 0.0948272ms / 1 = 0.0948272ms, 4%
hip::hip_copy_literal<float_type()>: 0.0864316ms / 100 = 0.000864316ms, 3%
gpu::code_object::mlir_convolution_add_relu<float_type(1x128x28x28:0x1x0x0, 1x512x28x28, 128x512x1x1, 1x128x28x28)>: 0.0857861ms / 3 = 0.0285954ms, 3%
gpu::code_object::add_relu_kernel<float_type(1x128x28x28, 1x128x28x28:0x1x0x0, 1x128x28x28)>: 0.0828137ms / 4 = 0.0207034ms, 3%
gpu::code_object::mlir_convolution_add_add_relu<float_type(1x512x28x28:0x1x0x0, 1x512x28x28, 1x128x28x28, 512x128x1x1, 1x512x28x28)>: 0.0796393ms / 3 = 0.0265464ms, 3%
gpu::convolution<float_type(1x256x28x28, 256x256x3x3, 3362816, 1x256x14x14)>: 0.073303ms / 1 = 0.073303ms, 3%
gpu::code_object::add_relu_kernel<float_type(1x512x7x7, 1x512x7x7:0x1x0x0, 1x512x7x7)>: 0.0717147ms / 3 = 0.0239049ms, 3%
gpu::code_object::add_relu_kernel<float_type(1x64x56x56, 1x64x56x56:0x1x0x0, 1x64x56x56)>: 0.0675753ms / 3 = 0.0225251ms, 3%
gpu::gemm<float_type(1x2048, 2048x1000, 1x1000, 1x1000)>: 0.0653694ms / 1 = 0.0653694ms, 3%
gpu::pooling<float_type(1x64x112x112, 1x64x56x56)>: 0.0635531ms / 1 = 0.0635531ms, 3%
load<float_type(9633792)>: 0.0614521ms / 71 = 0.000865522ms, 3%
gpu::code_object::mlir_convolution_add_add_relu<float_type(1x256x56x56:0x1x0x0, 1x256x56x56, 1x64x56x56, 256x64x1x1, 1x256x56x56)>: 0.0602779ms / 2 = 0.030139ms, 2%
gpu::code_object::mlir_convolution_add_relu<float_type(1x2048x7x7:0x1x0x0, 1x1536x7x7, 2048x1536x1x1, 1x2048x7x7)>: 0.0601698ms / 1 = 0.0601698ms, 2%
gpu::convolution<float_type(1x128x56x56, 128x128x3x3, 0, 1x128x28x28)>: 0.0594014ms / 1 = 0.0594014ms, 2%
gpu::code_object::mlir_convolution_add_relu<float_type(1x64x56x56:0x1x0x0, 1x256x56x56, 64x256x1x1, 1x64x56x56)>: 0.0504018ms / 2 = 0.0252009ms, 2%
gpu::code_object::mlir_convolution_add_relu<float_type(1x1024x14x14:0x1x0x0, 1x768x14x14, 1024x768x1x1, 1x1024x14x14)>: 0.0413323ms / 1 = 0.0413323ms, 2%
gpu::code_object::mlir_convolution_add_relu<float_type(1x512x14x14:0x1x0x0, 1x1024x14x14, 512x1024x1x1, 1x512x14x14)>: 0.0411509ms / 1 = 0.0411509ms, 2%
gpu::code_object::mlir_convolution_add_relu<float_type(1x64x112x112:0x1x0x0, 1x3x224x224, 64x3x7x7, 1x64x112x112)>: 0.0409082ms / 1 = 0.0409082ms, 2%
gpu::code_object::mlir_convolution_add_add_relu<float_type(1x2048x7x7:0x1x0x0, 1x2048x7x7, 1x512x7x7, 2048x512x1x1, 1x2048x7x7)>: 0.0344566ms / 1 = 0.0344566ms, 2%
gpu::code_object::mlir_convolution_add_relu<float_type(1x512x28x28:0x1x0x0, 1x384x28x28, 512x384x1x1, 1x512x28x28)>: 0.0341658ms / 1 = 0.0341658ms, 2%
gpu::code_object::mlir_convolution<float_type(1x512x7x7, 2048x512x1x1, 1x2048x7x7)>: 0.0330868ms / 1 = 0.0330868ms, 2%
gpu::code_object::mlir_convolution_add_relu<float_type(1x256x56x56:0x1x0x0, 1x128x56x56, 256x128x1x1, 1x256x56x56)>: 0.0306584ms / 1 = 0.0306584ms, 2%
gpu::code_object::mlir_convolution_add_relu<float_type(1x256x28x28:0x1x0x0, 1x512x28x28, 256x512x1x1, 1x256x28x28)>: 0.0302632ms / 1 = 0.0302632ms, 2%
gpu::code_object::mlir_convolution_add_relu<float_type(1x128x56x56:0x1x0x0, 1x256x56x56, 128x256x1x1, 1x128x56x56)>: 0.0294495ms / 1 = 0.0294495ms, 1%
gpu::code_object::mlir_convolution_add_relu<float_type(1x64x56x56:0x1x0x0, 1x64x56x56, 64x64x1x1, 1x64x56x56)>: 0.025836ms / 1 = 0.025836ms, 1%
gpu::code_object::concat_kernel<float_type(1x128x28x28, 1x256x28x28:802816x3136x112x2, 1x384x28x28)>: 0.0218758ms / 1 = 0.0218758ms, 1%
gpu::code_object::concat_kernel<float_type(1x64x56x56, 1x64x56x56, 1x128x56x56)>: 0.0208962ms / 1 = 0.0208962ms, 1%
gpu::code_object::concat_kernel<float_type(1x256x14x14, 1x512x14x14:401408x784x56x2, 1x768x14x14)>: 0.0208275ms / 1 = 0.0208275ms, 1%
gpu::code_object::concat_kernel<float_type(1x512x7x7, 1x1024x7x7:200704x196x28x2, 1x1536x7x7)>: 0.0205127ms / 1 = 0.0205127ms, 1%
broadcast<float_type(256)>: 0.0204408ms / 15 = 0.00136272ms, 1%
gpu::code_object::add_add_relu_reduce_mean_kernel<float_type(1x2048x7x7, 1x2048x7x7:0x1x0x0, 1x2048x7x7, 1x2048x1x1)>: 0.0197452ms / 1 = 0.0197452ms, 1%
load<int8_type(9633792)>: 0.0128907ms / 16 = 0.000805669ms, 1%
broadcast<float_type(512)>: 0.0116381ms / 10 = 0.00116381ms, 1%
broadcast<float_type(128)>: 0.0104227ms / 8 = 0.00130284ms, 1%
broadcast<float_type(64)>: 0.0082015ms / 7 = 0.00117164ms, 1%
broadcast<float_type(1024)>: 0.00690504ms / 6 = 0.00115084ms, 1%
broadcast<float_type(2048)>: 0.0037036ms / 3 = 0.00123453ms, 1%
@param<float_type()>: 0.00172414ms / 2 = 0.00086207ms, 1%
step<float_type(1x1024x14x14)>: 0.001618ms / 1 = 0.001618ms, 1%
step<float_type(1x256x56x56)>: 0.00146156ms / 1 = 0.00146156ms, 1%
step<float_type(1x512x28x28)>: 0.00142664ms / 1 = 0.00142664ms, 1%
reshape_lazy<float_type(1x2048x1x1)>: 0.00134836ms / 1 = 0.00134836ms, 1%
hip::hip_allocate_memory<int8_type()>: 0.00103484ms / 1 = 0.00103484ms, 1%
check_context::migraphx::gpu::context<float_type()>: 0.0009883ms / 1 = 0.0009883ms, 1%
```