@giuseros 2024-02-12
Hi @pfultz2 , just to be 100% sure, this 
```
gpu::convolution<float_type(1x256x14x14, 256x256x3x3, 2760704, 1x256x14x14)>
```
Is a `conv(A,B)` with `A=1x256x14x14, B=256x256x3x3`, right?